fix: API 변경사항 반영
